Image Photo progress bi!@#es

+++ DataReformat/PhotoExtractor.py
@@ -9,7 +9,7 @@ Andrew Neufeld, and Anthony Sopkow.
 
 import sys, os, shutil, tarfile, ntpath, zipfile, re, codecs, xml.etree.ElementTree as ET
 from msvcrt import getch
-from PIL import Image, ImageOps
+from PIL import Image, ImageOps, ImageDraw, ImageFont
 
 print ("Photo Extraction Script")
 
@@ -110,15 +110,17 @@ def convertFilesRecursively():
             #This is where we actually create the .zip files
             if files.endswith(".docx"):
                 try:
-                    #print("extracting file: " + os.path.join(r, files))
+                    print("extracting file: " + os.path.join(r, files))
                     if extractFilesFromDoc(files, r):
                         parseXmlToRenameImages(files, r)
                         #Done with the 'word' directory, remove it
                         shutil.rmtree(filename + "/word")
                 except UnicodeEncodeError:
-                    print ("Error with converting file: " + str.encode(files), file=logFile)
+                    #will have to change this back to r
+                    print ("Error with converting file: " + files, file=logFile)
                 except OSError:
-                    print ("Error with converting file: " + str.encode(files), file=logFile)
+                    #will have to change this back to r
+                    print ("Error with converting file: " + files, file=logFile)
 
 def ExtractPhotos():
     print("PhotoExtractor should only be run once for each folder structure and can increase disk usage. Continue? (y/n)")
@@ -143,26 +145,41 @@ def extractPilLibrary():
         print("PIL 1.1.7 library already exists")
 
 def processImageForSifte(rootPath, folderName, imageTitle, extension):
-    completeImageName = os.path.join(os.path.join(rootPath, folderName), imageTitle)
-    image = Image.open(completeImageName + extension)
-    #scale image
-    imageWidth, imageHeight = image.size
-    ratio = min(128/imageWidth, 98/imageHeight)
-    newImageHeight = int(imageHeight*ratio)
-    newImageWidth = int(imageWidth*ratio)
-    size = newImageWidth, newImageHeight
-    #image = image.resize(size)
-    #add white bar to bottom for text
-    #image = ImageOps.fit(image, (128, 128), Image.NEAREST, 0, (1.0, 0.0))
-    image = image.crop((0,0,newImageWidth,128))
-    
-    #draw.text((10, 103), )
-    #convert to png
-    image.save(completeImageName + ".png")
-    #remove old image
-    if not extension == ".png":
-        os.remove(completeImageName + extension)
-    
+    if not imageTitle == '':
+        completeImageName = os.path.join(os.path.join(rootPath, folderName), imageTitle)
+        image = Image.open(completeImageName + extension)
+        #scale image
+        imageWidth, imageHeight = image.size
+        ratio = min(128/imageWidth, 98/imageHeight)
+        newImageHeight = int(imageHeight*ratio)
+        newImageWidth = int(imageWidth*ratio)
+        size = newImageWidth, newImageHeight
+        image = image.resize(size)
+        #image2 = image.thumbnail(size, Image.ANTIALIAS)
+        #add white bar to bottom for text
+        #image = ImageOps.fit(image, (128, 128), Image.NEAREST, 0, (1.0, 0.0))
+        horizontalImageSpacing = int((128-newImageWidth)/2)
+        image = image.crop(( -1*horizontalImageSpacing,0,horizontalImageSpacing,128))
+        draw = ImageDraw.Draw(image)
+        myFont = ImageFont.truetype("Arial.ttf", 18, encoding="unic")
+        textWidth, textHeight = myFont.getsize(imageTitle)
+        textHorizontalPosition = (newImageWidth/2) - (textWidth/2)
+        textVerticalPosition = 98 + (15 - (textHeight/2))
+        #myFont = ImageFont.load("timR14.pil")
+        #border
+        draw.text((textHorizontalPosition-1, textVerticalPosition), imageTitle, fill='white', font=myFont)
+        draw.text((textHorizontalPosition+1, textVerticalPosition), imageTitle, fill='white', font=myFont)
+        draw.text((textHorizontalPosition, textVerticalPosition-1), imageTitle, fill='white', font=myFont)
+        draw.text((textHorizontalPosition, textVerticalPosition+1), imageTitle, fill='white', font=myFont)
+        draw.text((textHorizontalPosition, textVerticalPosition), imageTitle, fill='black', font=myFont)
+        #draw.text((10, 103), )
+        #convert to png
+        image.save(completeImageName + ".png", quality=100)
+        #image2.save(completeImageName + "2.png", quality=100)
+        #remove old image
+        if not extension == ".png":
+            os.remove(completeImageName + extension)
+        
 
 
 #extractPilLibrary()
